requirements.txt
clickhouse_driver==0.2.8
pandas==2.2.2
PyYAML==6.0.1
infi-clickhouse_orm==2.1.3
python-dotenv==1.0.1
types-pyyaml==6.0.12.11
docopt==0.6.2
